Fix for messed up packages Packages were messed up due to a checkstyle rename earlier.
