+++ src/main/java/spaceinvaders/group_22/ui/package-info.java
@@ -5,4 +5,4 @@
  * @author Jochem
  *
  */
-package ui;
+package spaceinvaders.group_22.ui;
