+++ src/main/resources/spaceinvaders/group_22/ui/package-info.java
@@ -1,0 +1,8 @@
+/**
+ * Contains resources for the ui package.
+ */
+/**
+ * @author Jochem
+ *
+ */
+package spaceinvaders.group_22.ui;
