+++ src/test/java/spaceinvaders/group_22/GameTest.java
@@ -1,4 +1,4 @@
-package group22.space_invaders;
+package spaceinvaders.group_22;
 
 import static org.junit.Assert.*;
