--- src/test/java/spaceinvaders/group_22/MainTest.java
@@ -1,22 +1,0 @@
-package spaceinvaders.group_22;
-
-import static org.junit.Assert.assertTrue;
-
-import org.junit.Test;
-
-/**
- * First test file for Main.java.
- * @author Jochem
- *
- */
-public class MainTest {
-
-	/**
-	 * Useless test to check if testing actually works.
-	 */
-	@Test
-	public final void test() {
-		assertTrue(true);
-	}
-
-}
